Move StrLitLexer from kore.syntax to kore.lexer
